move back ClassPathIntrospector to the plugin, and other re-factorings
