+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/NodeUpdateAbstractMojo.java
@@ -38,7 +38,7 @@ import org.reflections.util.ConfigurationBuilder;
 import com.vaadin.flow.component.dependency.HtmlImport;
 import com.vaadin.flow.plugin.common.FlowPluginFileUtils;
 import com.vaadin.flow.server.Constants;
-import com.vaadin.flow.server.frontend.ClassPathIntrospector.ClassFinder;
+import com.vaadin.flow.server.frontend.ClassFinder;
 import com.vaadin.flow.server.frontend.FrontendUtils;
 import com.vaadin.flow.server.frontend.NodeUpdater;
