+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/NodeUpdatePackages.java
@@ -33,7 +33,6 @@ import java.util.stream.Collectors;
 import org.apache.commons.io.FileUtils;
 
 import com.vaadin.flow.component.dependency.NpmPackage;
-import com.vaadin.flow.server.frontend.ClassPathIntrospector.ClassFinder;
 
 import elemental.json.Json;
 import elemental.json.JsonObject;
