+++ flow-test-generic/src/main/java/com/vaadin/flow/testutil/ClassesSerializableTest.java
@@ -152,7 +152,7 @@ public abstract class ClassesSerializableTest {
                 "com\\.vaadin\\.flow\\.component\\.internal\\.HtmlImportParser",
                 "com\\.vaadin\\.flow\\.server\\.webcomponent\\.WebComponentGenerator",
                 "com\\.vaadin\\.flow\\.server\\.communication\\.WebComponentBootstrapHandler(\\$.*)?",
-                "com\\.vaadin\\.flow\\.server\\.frontend\\.FlowClassVisitor(\\$.*)?",
+                "com\\.vaadin\\.flow\\.server\\.frontend\\.FrontendClassVisitor(\\$.*)?",
 
                 //Various test classes
                 ".*\\.test(s)?\\..*",
